requirements.txt
fastapi>=0.117.1
uvicorn>=0.37.0
pandas>=2.3.2
python-multipart>=0.0.20
openpyxl>=3.1.5
python-dateutil>=2.9.0
pytz>=2025.2
sqlalchemy>=2.0.23
pytest>=7.4.3
python-Levenshtein>=0.23.0
numpy>=1.26.1